Actual commit

+++ const.h
@@ -1,1 +1,1 @@
-const float PI = 3.14159265;
+const double PI = 3.14159265;
